src/test/java/ch/uzh/ifi/hase/soprafs23/service/LoginServiceTest.java
package ch.uzh.ifi.hase.soprafs23.service;

import ch.uzh.ifi.hase.soprafs23.constant.UserStatus;
import ch.uzh.ifi.hase.soprafs23.entity.Login;
import ch.uzh.ifi.hase.soprafs23.entity.User;
import ch.uzh.ifi.hase.soprafs23.repository.LoginRepository;
import ch.uzh.ifi.hase.soprafs23.repository.UserRepository;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.mockito.MockitoAnnotations;
import org.springframework.web.server.ResponseStatusException;

import java.util.ArrayList;
import java.util.Date;
import java.util.List;

import static org.junit.jupiter.api.Assertions.*;

public class LoginServiceTest {

/*TODO: Fix the tests for LoginService, the error is "Login failed because username does not exist or password does not match.", it needs to be adjusted */
/*
    @Mock
    private LoginRepository loginRepository;

    @Mock
    private UserRepository userRepository;

    @InjectMocks
    private LoginService loginService;

    private User testUser;
    private Login testLogin;
    private Login testLogin2;

    @BeforeEach
  public void setup() {
        MockitoAnnotations.openMocks(this);

        // create a test user and add it to userRepository
        testUser = new User();
        testUser.setUsername("testUsername");
        testUser.setPassword("testPassword");
        testUser.setUserId(1L);
        testUser.setToken("1");
        testUser.setCreation_date(new Date());
        testUser.setUserStatus(UserStatus.OFFLINE);
        userRepository.save(testUser);

        Mockito.when(userRepository.save(Mockito.any())).thenReturn(testUser);
        Mockito.when(userRepository.findByUserId(1L)).thenReturn(testUser);

        // create test login objects
        testLogin = new Login();
        testLogin.setUsername(testUser.getUsername());
        testLogin.setPassword(testUser.getPassword());
        testLogin.setUserId(1L);
        testLogin.setToken("1");
        testLogin.setLoginDate(new Date());

        testLogin2 = new Login();
        testLogin2.setUsername("testUsername2");
        testLogin2.setPassword("testPassword2");
        testLogin2.setUserId(2L);
        testLogin2.setToken("2");
        testLogin2.setLoginDate(new Date());

        Mockito.when(loginRepository.save(Mockito.any())).thenReturn(testLogin);
        Mockito.when(loginRepository.findByUserId(1L)).thenReturn(testLogin);

    }

    @Test
    public void createLogin_validInputs_success() {
        // when -> any object is being save in loginRepository -> return the dummy
        // testLogin
        Login createdLogin = loginService.createLogin(testLogin);

        // then
        Mockito.verify(loginRepository, Mockito.times(1)).save(Mockito.any());

        assertEquals(testLogin.getUserId(), createdLogin.getUserId());
        assertEquals(testLogin.getPassword(), createdLogin.getPassword());
        assertEquals(testLogin.getUsername(), createdLogin.getUsername());
        assertNotNull(createdLogin.getToken());
    }

    @Test
    public void createLogin_duplicateUsername_throwsException() {
        // given -> a first user has already been created
        loginService.createLogin(testLogin);

        // when -> setup additional mocks for LoginRepository
        //Mockito.when(loginRepository.findByPassword(Mockito.any())).thenReturn(testUser);
        Mockito.when(loginRepository.findByUsername(Mockito.any())).thenReturn(testLogin);

        // then -> attempt to create second user with same user -> check that an error
        // is thrown
        assertThrows(ResponseStatusException.class, () -> loginService.createLogin(testLogin));
    }

    @Test
    public void checkingUserSuccess() {
        Mockito.when(loginRepository.findByUsername(Mockito.anyString())).thenReturn(testLogin);

        Boolean returnedBool = loginService.checkPassword(testLogin);

        assertEquals(returnedBool, true);
    }

    @Test
    public void checkingUserFailureNull() {
        Mockito.when(loginRepository.findByUsername(Mockito.anyString())).thenReturn(null);

        assertThrows(ResponseStatusException.class, () -> loginService.checkPassword(testLogin));
    }

    @Test
    public void checkingUserFailureWrongPassword() {
        Mockito.when(userRepository.findByUsername(Mockito.anyString())).thenReturn(testUser);

        assertThrows(ResponseStatusException.class, () -> loginService.checkPassword(testLogin2));
    }

*/
}
